@@ -17,9 +17,9 @@
     "d3-tip": "^0.9.1",
     "history": "^4.7.2",
     "material-table": "^1.14.0",
-    "react": "^16.6.0",
+    "react": "^16.8.1",
     "react-apexcharts": "^1.0.10",
-    "react-dom": "^16.6.0",
+    "react-dom": "^16.8.1",
     "react-router-dom": "^4.3.1",
     "react-scripts": "^2.1.3",
     "react-vis-force": "^0.3.1",
