@@ -1,4 +1,4 @@
-import React, { Component, Fragment } from 'react'
+import React, { Fragment, useState } from 'react'
 import PropTypes from 'prop-types'
 import { NavLink } from 'react-router-dom'
 import { withStyles } from '@material-ui/core/styles'
@@ -26,76 +26,70 @@ const styles = (theme) => ({
     },
 })
 
-class sideBar extends Component {
-    state = {
-        open: false,
-    }
+const SideBar = (props) => {
+    const [open, setOpen] = useState(false)
+    const { classes, menuItems } = props
 
-    clickHandler = () => {
-        this.setState(state => ({ open: !state.open }))
-    }
+    const clickHandler = () => { setOpen(!open) }
 
-    render() {
-        const { classes, menuItems } = this.props
-        return (
-            <div className={ classes.sideBar }>
-                {
-                    Object.keys(menuItems).map((key) => {
-                        let submenu = menuItems[key]
-                        return (
-                            <Fragment key={ key }>
-                                <List>
-                                    {
-                                        submenu.items.map((item) => {
-                                            return item.hasOwnProperty('submenu') ? (
-                                                <Fragment key={ item.href }>
-                                                    <ListItem button onClick={ this.clickHandler } key={ item.href } className={ classes.itemLink }>
-                                                        <ListItemIcon className={ classes.itemIcon }>{ item.icon }</ListItemIcon>
-                                                        <ListItemText className={ classes.itemText } primary={ item.text } />
-                                                        { this.state.open ? <ExpandLessIcon /> : <ExpandMoreIcon /> }
-                                                    </ListItem>
-                                                    <Collapse in={ this.state.open } timeout="auto" unmountOnExit>
-                                                        <List component="div">
-                                                            {
-                                                                item.submenu.map((item) => {
-                                                                    return (
-                                                                        <ListItem button component={ NavLink } to={ item.path } exact
-                                                                            key={ item.text + item.path } className={ classes.itemLink } activeClassName={ classes.activeLink }
-                                                                        >
-                                                                            <ListItemIcon className={ classes.itemIcon }>{ item.icon }</ListItemIcon>
-                                                                            <ListItemText className={ classes.itemText } primary={ item.text } />
-                                                                        </ListItem>
-                                                                    )
-                                                                })
-                                                            }
-                                                        </List>
-                                                    </Collapse>
-                                                </Fragment>
-                                            ) : (
-                                                <ListItem component={ NavLink } to={ item.href } exact
-                                                    button key={ item.text } className={ classes.itemLink }
-                                                    activeClassName={ classes.activeLink } disabled={ item.disabled }
-                                                >
+    return (
+        <div className={ classes.sideBar }>
+            {
+                Object.keys(menuItems).map((key) => {
+                    let submenu = menuItems[key]
+                    return (
+                        <Fragment key={ key }>
+                            <List>
+                                {
+                                    submenu.items.map((item) => {
+                                        return item.hasOwnProperty('submenu') ? (
+                                            <Fragment key={ item.href }>
+                                                <ListItem button onClick={ clickHandler } key={ item.href } className={ classes.itemLink }>
                                                     <ListItemIcon className={ classes.itemIcon }>{ item.icon }</ListItemIcon>
                                                     <ListItemText className={ classes.itemText } primary={ item.text } />
+                                                    { open ? <ExpandLessIcon /> : <ExpandMoreIcon /> }
                                                 </ListItem>
-                                            )
-                                        })
-                                    }
-                                </List>
-                                <Divider />
-                            </Fragment>
-                        )
-                    })
-                }
-                <Divider />
-            </div>
-        )
-    }
+                                                <Collapse in={ open } timeout="auto" unmountOnExit>
+                                                    <List component="div">
+                                                        {
+                                                            item.submenu.map((item) => {
+                                                                return (
+                                                                    <ListItem button component={ NavLink } to={ item.path } exact
+                                                                        key={ item.text + item.path } className={ classes.itemLink } activeClassName={ classes.activeLink }
+                                                                    >
+                                                                        <ListItemIcon className={ classes.itemIcon }>{ item.icon }</ListItemIcon>
+                                                                        <ListItemText className={ classes.itemText } primary={ item.text } />
+                                                                    </ListItem>
+                                                                )
+                                                            })
+                                                        }
+                                                    </List>
+                                                </Collapse>
+                                            </Fragment>
+                                        ) : (
+                                            <ListItem component={ NavLink } to={ item.href } exact
+                                                button key={ item.text } className={ classes.itemLink }
+                                                activeClassName={ classes.activeLink } disabled={ item.disabled }
+                                            >
+                                                <ListItemIcon className={ classes.itemIcon }>{ item.icon }</ListItemIcon>
+                                                <ListItemText className={ classes.itemText } primary={ item.text } />
+                                            </ListItem>
+                                        )
+                                    })
+                                }
+                            </List>
+                            <Divider />
+                        </Fragment>
+                    )
+                })
+            }
+            <Divider />
+        </div>
+    )
 }
 
-sideBar.propTypes = {
+SideBar.propTypes = {
     menuItems: PropTypes.array,
 }
 
-export default withStyles(styles)(sideBar)
+export default withStyles(styles)(SideBar)
