@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React, { useState, useEffect } from 'react'
 import axios from 'axios'
 
 import Heading from '../../components/Typography/Heading'
@@ -8,39 +8,27 @@ import { CircularLoader } from '../../components/Progress/Progress'
 
 const stagesUrl = process.env.NODE_ENV === 'production' ? 'https://pmd.renci.org/api/proposals/by-stage' : 'http://localhost:3030/proposals/by-stage'
 
-class proposalsByStage extends Component {
-    state = {
-        stages: [],
-    }
+const proposalsByStage = (props) => {
+    const [stages, setStages] = useState([])
     
-    async fetchProposals() {
-        await axios.get(stagesUrl)
-            .then(response => {
-                this.setState({ stages: response.data, })
-            })
-            .catch(error => {
-                console.error('Error:', error)
-            })
-    }
+    useEffect(() => {
+        axios.get(stagesUrl)
+            .then(response => setStages(response.data))
+            .catch(error => console.error('Error:', error))
+    }, [])
 
-    componentDidMount = this.fetchProposals
-
-    render() {
-        const { stages } = this.state
-        return (
-            <div>
-                <Heading>
-                    Proposals by Stage
-                </Heading>
-                
-                {
-                    stages.length > 0
-                    ? <ProposalsByStageTable proposalsByStage={ this.state.stages }/>
-                    : <CircularLoader/>
-                }
-            </div>
-        )
-    }
+    return (
+        <div>
+            <Heading>
+                Proposals by Stage
+            </Heading>
+            {
+                stages.length > 0
+                ? <ProposalsByStageTable proposalsByStage={ stages }/>
+                : <CircularLoader/>
+            }
+        </div>
+    )
 }
 
 export default proposalsByStage
