@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React, { useState, useEffect } from 'react'
 import axios from 'axios'
 import Heading from '../../components/Typography/Heading'
 
@@ -8,42 +8,33 @@ import { CircularLoader } from '../../components/Progress/Progress'
 const proposalsUrl = process.env.NODE_ENV === 'production' ? 'https://pmd.renci.org/api/proposals/submitted-services' : 'http://localhost:3030/proposals/submitted-services'
 const servicesUrl = process.env.NODE_ENV === 'production' ? 'https://pmd.renci.org/api/services/submission' : 'http://localhost:3030/services/submission'
 
-class proposalsByApproval extends Component {
-    state = {
-        proposals: [],
-        services: [],
-    }
+const ProposalsByApproval = (props) => {
+    const [proposals, setProposals] = useState([])
+    const [services, setServices] = useState([])
 
-    componentDidMount() {
+    useEffect(() => {
         const proposalsPromise = axios.get(proposalsUrl)
         const servicesPromise = axios.get(servicesUrl)
         Promise.all([proposalsPromise, servicesPromise])
             .then((response) => {
-                this.setState({
-                    proposals: response[0].data,
-                    services: response[1].data,
-                })
+                setProposals(response[0].data)
+                setServices(response[1].data)
             })
-            .catch(error => {
-                console.error('Error:', error)
-            })
-    }
+            .catch(error => console.error('Error:', error))
+    }, [])
 
-    render() {
-        const { proposals, services } = this.state
-        return (
-            <div>
-                <Heading>Submitted Proposals</Heading>
-                
-                {
-                    (proposals.length > 0 && services.length > 0)
-                    ? <ProposalsMatrix proposals={ proposals } services={ services }/>
-                    : <CircularLoader/>
-                }
-                
-            </div>
-        )
-    }
+    return (
+        <div>
+            <Heading>Submitted Proposals</Heading>
+            
+            {
+                (proposals.length > 0 && services.length > 0)
+                ? <ProposalsMatrix proposals={ proposals } services={ services }/>
+                : <CircularLoader/>
+            }
+            
+        </div>
+    )
 }
 
-export default proposalsByApproval
+export default ProposalsByApproval
